- minor refactoring: moved moved internal classes to internal structure and removed unnessessary interfaces
