+++ Sdl.CommunityToolkit.FileType/Processors/Internal/Comparers/RegexSearchResultsComparer.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections.Generic;
+
+namespace Sdl.Community.Toolkit.FileType.Processors.Internal.Comparers
+{
+	internal class RegexSearchResultsComparer : IComparer<RegexMatch>
+	{
+		public int Compare(RegexMatch x, RegexMatch y)
+		{
+			if (x == null || y == null)
+			{
+				return 0;
+			}
+
+			return x.Index.CompareTo(y.Index);
+		}
+	}
+}
