+++ Sdl.CommunityToolkit.FileType/Processors/Internal/RegexMatch.cs
@@ -1,0 +1,36 @@
+﻿using Sdl.Community.Toolkit.FileType.Processors.API;
+
+namespace Sdl.Community.Toolkit.FileType.Processors.Internal
+{
+	internal class RegexMatch
+	{
+		public enum TagType
+		{
+			Placeholder, TagPairOpening, TagPairClosing
+		}
+
+		public IMatchRule Rule
+		{
+			get;
+			set;
+		}
+
+		public int Index
+		{
+			get;
+			set;
+		}
+
+		public string Value
+		{
+			get;
+			set;
+		}
+
+		public TagType Type
+		{
+			get;
+			set;
+		}		
+	}
+}
