+++ Sdl.CommunityToolkit.FileType/Sdl.Community.Toolkit.FileType.csproj
@@ -76,16 +76,15 @@
     <Compile Include="OriginType.cs" />
     <Compile Include="Processors\API\IEmbeddedContentProcessorSettings.cs" />
     <Compile Include="Processors\API\IMatchRule.cs" />
-    <Compile Include="Processors\API\IRegexMatch.cs" />
-    <Compile Include="Processors\Comparers\MatchRulesTypeComparer.cs" />
-    <Compile Include="Processors\Comparers\RegexSearchResultsComparer.cs" />
-    <Compile Include="Processors\EmbeddedContentConstants.cs" />
+    <Compile Include="Processors\Internal\Comparers\MatchRulesTypeComparer.cs" />
+    <Compile Include="Processors\Internal\Comparers\RegexSearchResultsComparer.cs" />
+    <Compile Include="Processors\Internal\EmbeddedContentConstants.cs" />
     <Compile Include="Processors\FeedbackRemovalProcessor.cs" />
     <Compile Include="Processors\Models\MatchRule.cs" />
-    <Compile Include="Processors\Models\RegexMatch.cs" />
+    <Compile Include="Processors\Internal\RegexMatch.cs" />
     <Compile Include="Processors\RegexEmbeddedNativeGenerationProcessor.cs" />
     <Compile Include="Processors\RegexEmbeddedNativeProcessor.cs" />
-    <Compile Include="Processors\Services\RegexProcessorService.cs" />
+    <Compile Include="Processors\Internal\Services\RegexProcessorService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Visitors\CommentDataVisitor.cs" />
     <Compile Include="Visitors\EndOfItemCharacterCounterNoTagsVisitor.cs" />
